fix: validate post request to check if user is allowed to edit a post

+++ src/main/java/com/example/demo/controller/PostRestController.java
@@ -4,29 +4,28 @@ import com.example.demo.dto.ApiResponse;
 import com.example.demo.dto.PostRequest;
 import com.example.demo.dto.PostResponse;
 import com.example.demo.mapper.Mapper;
-import com.example.demo.model.Post;
+import com.example.demo.entity.Post;
+import com.example.demo.service.PostPermissionService;
 import com.example.demo.service.interfaces.PostService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
+import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
 @RestController
 @RequestMapping("${api.base-url}/posts")
+@AllArgsConstructor
 @Tag(name = "Post Management", description = "Operations for creating, searching, and managing blog posts")
 public class PostRestController {
     private final PostService postService;
+    private final PostPermissionService postPermissionService;
     private final Mapper<Post, PostResponse, PostRequest> postMapper;
-
-    public PostRestController(PostService postService, Mapper<Post, PostResponse, PostRequest> postMapper) {
-        this.postMapper = postMapper;
-        this.postService = postService;
-    }
 
     @Operation(summary = "Get all posts", description = "Retrieves a full list of all blog posts from the system.")
     @ApiResponses(value = {
@@ -48,7 +47,6 @@ public class PostRestController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "404", description = "Post not found")
     })
     @GetMapping("/{postId}")
-    @PreAuthorize("hasRole('ADMIN')")
     public ApiResponse<PostResponse> getById(@Valid @PathVariable String postId) {
         return new ApiResponse<>(HttpStatus.OK, "Post retrieved successfully", postMapper.toResponse(postService.getById(postId)));
     }
@@ -58,8 +56,10 @@ public class PostRestController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "201", description = "Post created successfully")
     })
     @PostMapping
-    public ApiResponse<PostResponse> create(@Valid @RequestBody PostRequest request) {
+    public ApiResponse<PostResponse> create(@Valid @RequestBody PostRequest request, Authentication auth) {
         var post = postMapper.toEntity(request);
+
+        post.setAuthorEmail(auth.getName());
 
         return new ApiResponse<>(HttpStatus.CREATED, "Post created successfully", postMapper.toResponse(postService.create(post)));
     }
@@ -69,8 +69,15 @@ public class PostRestController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "200", description = "Post deleted successfully")
     })
     @DeleteMapping("/{postId}")
-    public ApiResponse<Void> delete(@Valid @PathVariable String postId) {
+    public ApiResponse<Void> delete(@PathVariable String postId, Authentication auth) {
+        String email = auth.getName();
+
+        if (!postPermissionService.canEditPost(postId, email)) {
+            throw new RuntimeException("User with email: %s does not have permission to delete post with id: %s".formatted(email, postId));
+        }
+
         postService.delete(postId);
+
         return new ApiResponse<>(HttpStatus.OK, "Post deleted successfully", null);
     }
 
@@ -80,7 +87,13 @@ public class PostRestController {
             @io.swagger.v3.oas.annotations.responses.ApiResponse(responseCode = "404", description = "Post not found")
     })
     @PatchMapping("/{postId}/content")
-    public ApiResponse<PostResponse> updatePostContent(@Valid @PathVariable String postId, @RequestBody PostRequest request) {
+    public ApiResponse<PostResponse> updatePostContent(@Valid @PathVariable String postId, @RequestBody PostRequest request, Authentication auth) {
+        String email = auth.getName();
+
+        if (!postPermissionService.canEditPost(postId, email)) {
+            throw new RuntimeException("User with email: %s does not have permission to edit post with id: %s".formatted(email, postId));
+        }
+
         var post = postService.updatePostContent(postId, request.getContent());
 
         return new ApiResponse<>(HttpStatus.OK, "Post content updated successfully", postMapper.toResponse(post));
